Fix Bast lion complex spawn, frog infest offset, and bast_curse flag clear.
Bast major raid now spawns from TEMPLE_COMPLEX_BAST like Seth/Ptah; frog last_infest uses uint16 so high grid offsets do not re-infest every tick; clear bast_curse_active only after a disease outcome.

+++ src/figuretype/figure_frog.cpp
@@ -204,9 +204,9 @@ void figure_frog::figure_action() {
         }
         {
             auto &d = runtime_data();
-            const int goff = tile().grid_offset();
+            const uint16_t goff = (uint16_t)tile().grid_offset();
             if (d.last_infest_offset != goff) {
-                d.last_infest_offset = (int16_t)goff;
+                d.last_infest_offset = goff;
                 try_infest_near(tile());
             }
         }
